+++ charts_plugin/src/main/java/com/futureworkshops/mobileworkflow/plugin/charts/pie/view/UIPieChartPluginStep.kt
@@ -6,11 +6,11 @@ package com.futureworkshops.mobileworkflow.plugin.charts.pie.view
 
 import com.futureworkshops.mobileworkflow.model.WorkflowServiceResponse
 import com.futureworkshops.mobileworkflow.plugin.charts.pie.step.PieChartItem
-import com.futureworkshops.mobileworkflow.surveykit.StepIdentifier
-import com.futureworkshops.mobileworkflow.surveykit.backend.views.step.StepView
-import com.futureworkshops.mobileworkflow.surveykit.result.StepResult
-import com.futureworkshops.mobileworkflow.surveykit.services.MobileWorkflowServices
-import com.futureworkshops.mobileworkflow.surveykit.steps.Step
+import com.futureworkshops.mobileworkflow.StepIdentifier
+import com.futureworkshops.mobileworkflow.backend.views.step.StepView
+import com.futureworkshops.mobileworkflow.result.StepResult
+import com.futureworkshops.mobileworkflow.services.MobileWorkflowServices
+import com.futureworkshops.mobileworkflow.steps.Step
 
 internal class UIPieChartPluginStep(
     private val title: String,
